@@ -4,8 +4,8 @@
   "license": "MIT",
   "scripts": {
     "ng": "ng",
-    "start": "ng serve --aot --proxy-config proxy.config.js --hmr -e=hmr",
-    "build:staging": "ng build --prod --env=staging",
+    "start": "ng serve --aot --proxy-config proxy.config.js --hmr",
+    "build:staging": "ng build --prod --configuration=staging",
     "build:prod": "ng build --prod",
     "test": "ng test",
     "lint": "ng lint",
@@ -13,42 +13,44 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "5.1.0",
-    "@angular/cdk": "5.0.1",
-    "@angular/common": "5.1.0",
-    "@angular/compiler": "5.1.0",
-    "@angular/core": "5.1.0",
-    "@angular/forms": "5.1.0",
-    "@angular/material": "5.0.1",
-    "@angular/platform-browser": "5.1.0",
-    "@angular/platform-browser-dynamic": "5.1.0",
-    "@angular/router": "5.1.0",
-    "angular-oauth2-oidc": "3.0.3",
-    "core-js": "2.5.3",
+    "@angular/animations": "6.0.3",
+    "@angular/cdk": "6.1.0",
+    "@angular/common": "6.0.3",
+    "@angular/compiler": "6.0.3",
+    "@angular/core": "6.0.3",
+    "@angular/forms": "6.0.3",
+    "@angular/material": "6.1.0",
+    "@angular/platform-browser": "6.0.3",
+    "@angular/platform-browser-dynamic": "6.0.3",
+    "@angular/router": "6.0.3",
+    "angular-oauth2-oidc": "4.0.2",
+    "core-js": "2.5.7",
     "hammerjs": "2.0.8",
-    "rxjs": "5.5.5",
+    "rxjs": "6.2.0",
+    "rxjs-compat": "6.2.0",
     "web-animations-js": "2.3.1",
-    "zone.js": "0.8.18"
+    "zone.js": "0.8.26"
   },
   "devDependencies": {
-    "@angular/cli": "1.6.0",
-    "@angular/compiler-cli": "5.1.0",
-    "@angular/language-service": "5.1.0",
-    "@types/jasmine": "2.8.2",
+    "@angular/cli": "6.0.5",
+    "@angular/compiler-cli": "6.0.3",
+    "@angular/language-service": "6.0.3",
+    "@angular-devkit/build-angular": "0.6.0",
+    "@types/jasmine": "2.8.7",
     "@types/jasminewd2": "2.0.3",
-    "@types/node": "8.0.58",
-    "codelyzer": "4.0.2",
-    "jasmine-core": "2.8.0",
+    "@types/node": "10.1.2",
+    "codelyzer": "4.3.0",
+    "jasmine-core": "3.1.0",
     "jasmine-spec-reporter": "4.2.1",
-    "karma": "1.7.1",
+    "karma": "2.0.2",
     "karma-chrome-launcher": "2.2.0",
     "karma-cli": "1.0.1",
-    "karma-coverage-istanbul-reporter": "1.3.0",
-    "karma-jasmine": "1.1.1",
-    "karma-jasmine-html-reporter": "0.2.2",
-    "protractor": "5.2.2",
-    "ts-node": "4.0.1",
-    "tslint": "5.8.0",
-    "typescript": "2.6.2"
+    "karma-coverage-istanbul-reporter": "2.0.1",
+    "karma-jasmine": "1.1.2",
+    "karma-jasmine-html-reporter": "1.1.0",
+    "protractor": "5.3.2",
+    "ts-node": "6.0.5",
+    "tslint": "5.10.0",
+    "typescript": "2.7.2"
   }
 }
