@@ -1,10 +1,7 @@
-﻿import { Injectable } from '@angular/core';
+import { Injectable } from '@angular/core';
 import { HttpHeaders } from '@angular/common/http';
 import { Router } from '@angular/router';
-import { Observable } from 'rxjs/Observable';
-import { BehaviorSubject } from 'rxjs/BehaviorSubject';
-import { interval } from 'rxjs/observable/interval';
-import { timer } from 'rxjs/observable/timer';
+import { Observable, BehaviorSubject, interval, timer } from 'rxjs';
 
 import { OAuthService } from 'angular-oauth2-oidc';
 
